+++ db/migrate/20140505165530_add_first_locked_at_to_sheets.rb
@@ -1,0 +1,5 @@
+class AddFirstLockedAtToSheets < ActiveRecord::Migration
+  def change
+    add_column :sheets, :first_locked_at, :datetime
+  end
+end