+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - Printed PDFs have better styling for sections and subsections
   - Moved project name to top center of PDF
   - Split subject name and subject acrostic across two lines on top left of PDF
+  - Sheet PDFs now display amendments made after a sheet is initially locked
 - **Gem Changes**
   - Updated to minitest-reporters 1.0.4